feat: whatwg url validation, improved ts support
BREAKING CHANGE: removed bluebird promises and validation methods are not bound to instance by
default. This may required apps performing the integration to change their code. http-url format
has been changed from ReDoS prone RegEx to whatwg url one. Validation is performed differently as a result
Opinionated format requires protocol to be http: or https: and port to either be omitted or be 80/443